remove legacy component pool
