+++ src/pages/Pools/components/ListPool/ListPool.tsx
@@ -3,7 +3,7 @@ import { Button } from 'components/Button';
 import { FallbackEmptyData } from 'components/FallbackEmptyData';
 import { Table, TableHeaderProps } from 'components/Table';
 import { formatDisplayUsdt, parseAssetOnlyDenom } from 'pages/Pools/helpers';
-import { PoolTableData } from 'pages/Pools/indexV3';
+import { PoolTableData } from 'pages/Pools';
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { PoolInfoResponse } from 'types/pool';
